fix(loans): integrate with book pages — clickable book title + Lend button
Book titles in the Loans tab are now Links to the book detail page (couldn't get there before). Books gain a "Lend this book" icon button next to the existing Edit / Refresh metadata actions, opening the loan modal pre-filled with that book.

Extracted LoanFormModal to src/components/ matching the existing AddEditionModal / EditBookModal pattern, and added a prefillBook prop so the book-page entry point can hide the search and pre-select.
